examples: change some pins to be usable on the STM32C0316-DK board
The STM32C0 board I have at the moment is the STM32C0316-DK "Discovery
kit with STM32C031C6 MCU" [1]. It's an inexpensive way to get going with
the STM32C031C6T6 chip and comes with a standalone STLinkV3-MINIE which
itself is valuable.

Update the button and uart examples in this repo to be usable on that
board by changing the pins around. PA5 is the user LED (active-low) on
the DK, which was already used in the blinky examples. clockout and rtic
still may need to be updated later.

[1] https://www.st.com/en/evaluation-tools/stm32c0316-dk.html