gpio: use proper field accessors rather than .bits()
Code is much more readable when we let the PAC name the register fields
and values for us, rather than doing bit twiddling by hand. The only
weird thing is the stray underscore in set_() for the BSRR fields. That
seems to be something to do with newer svd2rust versions but I can't
find a clear indication in the changelog.

This removes all uses of .bits except for the external interrupt
handling, because those registers are weird and don't have array style
accessors yet.